Tooling: Provisioning scripts and docs for local development
This expands the provisioning scripts a bit to include additional
plugins used by the News site, configuring some WP settings, and
importing demo content. The content comes from the actual News site, but
it has been scrubbed of PII and there is a readme describing how to do
the scrub again if we want to update the demo content in the future.

It also adds more details in the main repo readme about how to set up
the dev environment.

Fixes #2