Diverge from Blockbase (#15)
Blockbase contains a lot of functionality that we don't need for the News site, particularly around custom controls and settings in the Customizer. This PR removes all of that, and also removes the blockbase_ prefix from functions in favor of a PHP namespace.

This does incorporate recent changes from Blockbase, however, that haven't been released on the theme directory yet, namely a different way of handling Google fonts.