Get fonts and headings styles to load correction on front/back
Fonts weren't loading because of a missing value in the query parameter
for the Google Fonts URL.

Media queries to adjust heading font sizes were being overridden in
the block editor (but not the front end), so this takes a queue from
Blockbase and defines some custom variables for headings typography
(even though the typography can be set in styles:elements). This way
the custom variables can have their values changed in the media query,
so it doesn't matter what order they load in.