Single post view (#21)
Adds block templates and template parts, tweaks the theme.json file, and adds some rules to the stylesheet to create the single post view.

This includes a responsive layout for the post title/meta area and the main post content column. The post content also accommodates wide and full alignment on individual blocks.

This doesn't include the global header or footer. It also doesn't include the social sharing links or like button shown in the mockup, because those elements are not yet available as site blocks. (There's an issue for the Like button, couldn't find one for social sharing icons.)

The mockup doesn't include a comment section, and it appears that the News site does not currently have comments enabled, so that has also been omitted.

Fixes #17