Add the Blue Bar (#41)
Introduces the "Blue Bar" or, more accurately, a local header beneath the global header. It includes a simple taxonomy-based breadcrumb on the left and a CSS-based dropdown menu of categories on the right. The color scheme changes ("The Not Blue Bar") for certain categories. The layout responsively handles mobile, tablet, and desktop views.

This also makes the entire site header (including the global header) fixed to the top when not in a narrow mobile viewport, as called for in the design.

Fixes #32