Add a wporg/post-title block to display the version numbers
Sorry if this steamrolls WIP @shaunandrews!

I added a custom post title block with a `releaseVersion` attribute to output just the version number. Feels like overkill but anything else I came up with was super hacky.

It seems to play ok with your CSS, I bodged it a bit to work. Not sure if other styling is needed.