Sort category menu by column instead of row (#277)
* Quick PoC

I can't figure out why the `flex-direction` in `break-small` seems to override the `column` sort. This breaks the vertical menu at small sizes, but has the correct sorting on wider screens.

* Fix column sorting.

* Hide child categories

Just in case - this will keep it neater if child categories get added to the site in future. Thanks @dd32!

* Try using text columns instead of flex

Co-authored-by: Kelly Dwan <ryelle@users.noreply.github.com>