Remove link from featured image in community category post grid items (#344)
Removing the link improves accessibility by eliminating a redundant
keypress when tabbing through the grid items. It does mean that less of
the grid item is mouse-clickable for navigating to an individual post,
but the post title is still prominent enough to provide an adequate
click target. Notably, this also doesn't affect the visual hover effects
when mousing over a grid item. This seems like a reasonable tradeoff to
improve a11y.

Fixes #340