Add styles for lists of "props"/"contributors" (#363)
Registers some block styles and adds related rules to the stylesheet to match the design for "Contributors" blocks. In some release posts (major releases), the [wpcredits] shortcode is used, which can't have the block styles applied to it, so there's a separate PR to add the correct classes to the container that it renders. For other release posts (minor releases), the list of props is usually generated via a script and pasted into a paragraph or code block. For these, the block style can be applied manually to be able to match the design.

Note that the design calls for the props block to be full width. It seemed to make more sense to just have the post author add the additional alignfull class to the block, rather than trying to match the same full width styles with another class name.

Fixes #355